feat: migrate to typescript
